final2

+++ Final/src/Keyword.java
@@ -1,10 +1,10 @@
 
 public class Keyword {
 	public String name;
-	public double weight;
+	public int weight;
 	public int count;
 	
-	public Keyword(String name,double weight){
+	public Keyword(String name,int weight){
 		this.name = name;
 		this.weight = weight;
 	}
